+++ src/modules/message/message.module.ts
@@ -3,6 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { MessageService } from './message.service';
 import { BulkMessageService } from './bulk-message.service';
 import { MessageTypeBackfillService } from './message-type-backfill.service';
+import { HistorySyncService } from './history-sync.service';
 import { MessageController } from './message.controller';
 import { SessionModule } from '../session/session.module';
 import { TemplateModule } from '../template/template.module';
@@ -12,7 +13,7 @@ import { MessageBatch } from './entities/message-batch.entity';
 @Module({
   imports: [TypeOrmModule.forFeature([Message, MessageBatch], 'data'), SessionModule, TemplateModule],
   controllers: [MessageController],
-  providers: [MessageService, BulkMessageService, MessageTypeBackfillService],
-  exports: [MessageService, BulkMessageService],
+  providers: [MessageService, BulkMessageService, MessageTypeBackfillService, HistorySyncService],
+  exports: [MessageService, BulkMessageService, HistorySyncService],
 })
 export class MessageModule {}